@robertnishihara 2019-05-26
@pcmoritz can you check if this PR affects performance?