@pcmoritz 2019-05-27
Looks like there is an about 10% regression for builder performance:

```
import pyarrow
o = pyarrow.serialize([1])
```

before:
```
In [3]: %timeit o.to_buffer()                                                                                                               
19.6 µs ± 261 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [4]: %timeit o.to_buffer()                                                                                                               
19.8 µs ± 831 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [5]: %timeit o.to_buffer()                                                                                                               
20 µs ± 668 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

after:
```
In [7]: %timeit o.to_buffer()                                                                                                               
22.2 µs ± 237 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [8]: %timeit o.to_buffer()                                                                                                               
22.1 µs ± 362 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [9]: %timeit o.to_buffer()                                                                                                               
22 µs ± 216 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```