@robertnishihara 2019-05-27
@pcmoritz I see, that's unfortunate. Did you also check end to end serialization and deserialization?